re ZON-3213: Combine the message config of `parse` & `urbanairship` into `mobile`.
- it's much cleaner to keep mobile settings in a single message config
  (contained settings are `mobile_enabled` and `mobile_text`)
- settings of both push messages were identical, since they were copied
- the new `mobile` message does not forward to a `mobile` push notifier, but
  calls the push notifier for `parse` and `urbanairship`
- the `mobile` message is also registered as `parse` for bw compat
