+++ src/zeit/push/browser/form.py
@@ -78,11 +78,7 @@ class SocialBase(zeit.cms.browser.form.CharlimitMixin):
             {'type': 'twitter',
              'enabled': data.pop('twitter_ressort_enabled', False),
              'account': data.pop('twitter_ressort', None)},
-            {'type': 'parse',
-             'enabled': data.get('mobile_enabled', False),
-             'override_text': data.get('mobile_text', None),
-             'channels': zeit.push.interfaces.PARSE_NEWS_CHANNEL},
-            {'type': 'urbanairship',
+            {'type': 'mobile',
              'enabled': data.pop('mobile_enabled', False),
              'override_text': data.pop('mobile_text', None),
              'channels': zeit.push.interfaces.PARSE_NEWS_CHANNEL},
