+++ src/zeit/push/parse.py
@@ -186,59 +186,6 @@ def from_product_config():
         int(config['parse-expire-interval']))
 
 
-class Message(zeit.push.message.Message):
-
-    grok.context(zeit.cms.content.interfaces.ICommonMetadata)
-    grok.name('parse')
-
-    @property
-    def text(self):
-        return self.context.title
-
-    @property
-    def additional_parameters(self):
-        result = {}
-        for name in ['teaserTitle', 'teaserText', 'teaserSupertitle']:
-            value = getattr(self.context, name)
-            if value:
-                result[name] = value
-        if self.image:
-            result['image_url'] = self.image.uniqueId.replace(
-                zeit.cms.interfaces.ID_NAMESPACE, 'http://images.zeit.de/')
-        return result
-
-    @property
-    def image(self):
-        images = zeit.content.image.interfaces.IImages(self.context, None)
-        if images is None or images.image is None:
-            return None
-        image = images.image
-        if zeit.content.image.interfaces.IImageGroup.providedBy(image):
-            for name in image:
-                if self._image_pattern in name:
-                    return image[name]
-        else:
-            return image
-
-    @property
-    def _image_pattern(self):
-        config = zope.app.appsetup.product.getProductConfiguration(
-            'zeit.push')
-        return config['parse-image-pattern']
-
-
-@grok.subscribe(
-    zeit.cms.content.interfaces.ICommonMetadata,
-    zeit.cms.checkout.interfaces.IBeforeCheckinEvent)
-def set_push_news_flag(context, event):
-    push = zeit.push.interfaces.IPushMessages(context)
-    for service in push.message_config:
-        if (service['type'] == 'parse' and service.get('enabled')
-                and service.get('channels') == PARSE_NEWS_CHANNEL):
-            context.push_news = True
-            break
-
-
 class PayloadDocumentation(Connection):
 
     def push(self, data):
